requirements.txt
alembic==1.8.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
Werkzeug==2.1.2
zipp==3.8.0
